+++ lib/flutter_screenguard_method_channel.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'dart:async';
 
 import 'flutter_screenguard_platform_interface.dart';
 
@@ -8,6 +9,48 @@ class MethodChannelFlutterScreenguard extends FlutterScreenguardPlatform {
   /// The method channel used to interact with the native platform.
   @visibleForTesting
   final methodChannel = const MethodChannel('flutter_screenguard');
+
+  final _screenshotChannel =
+      const MethodChannel('flutter_screenguard_screenshot_event');
+  final _recordingChannel =
+      const MethodChannel('flutter_screenguard_screen_recording_event');
+
+  final _screenshotController =
+      StreamController<Map<String, dynamic>>.broadcast();
+  final _recordingController =
+      StreamController<Map<String, dynamic>>.broadcast();
+
+  bool _handlersInitialized = false;
+
+  void _ensureHandlersInitialized() {
+    if (_handlersInitialized) return;
+    _handlersInitialized = true;
+
+    _screenshotChannel.setMethodCallHandler((call) async {
+      if (call.method == 'onScreenshotCaptured') {
+        _screenshotController.add(Map<String, dynamic>.from(call.arguments));
+      }
+    });
+
+    _recordingChannel.setMethodCallHandler((call) async {
+      if (call.method == 'onScreenRecordingCaptured') {
+        _recordingController
+            .add(Map<String, dynamic>.from(call.arguments ?? {}));
+      }
+    });
+  }
+
+  @override
+  Stream<Map<String, dynamic>> get onScreenshotCaptured {
+    _ensureHandlersInitialized();
+    return _screenshotController.stream;
+  }
+
+  @override
+  Stream<Map<String, dynamic>> get onScreenRecordingCaptured {
+    _ensureHandlersInitialized();
+    return _recordingController.stream;
+  }
 
   static const List<Alignment> alignments = [
     Alignment.topLeft,
@@ -24,26 +67,16 @@ class MethodChannelFlutterScreenguard extends FlutterScreenguardPlatform {
   /// activate a screenshot blocking with a color effect view (iOS 13+, Android 8+)
   /// [color] color of the background
   ///
-  /// [timeAfterResume] (Android only) Time delayed for the view to stop displaying when going back
-  /// to the application (in milliseconds). Default = 1000ms
-  ///
-  /// function will throw warning when [timeAfterResume] bigger than 3000ms,
-  /// users have to wait for the application to turn off the filter before going back
-  /// to the main view, which is a very bad user experiences.
-  ///
   /// Throws a [PlatformException] if there were technical problems on native side
   /// (e.g. lack of relevant hardware).
   @override
   Future<void> register({
     required Color color,
-    Duration? timeAfterResume,
   }) async {
     final colorHex =
-        '#${color.value.toRadixString(16).padLeft(8, '0').substring(2)}';
-    await methodChannel.invokeMethod<void>('register', <String, dynamic>{
+        '#${color.toARGB32().toRadixString(16).padLeft(8, '0').substring(2)}';
+    await methodChannel.invokeMethod<void>('activateShield', <String, dynamic>{
       'color': colorHex,
-      'timeAfterResume': (timeAfterResume ?? const Duration(milliseconds: 1000))
-          .inMilliseconds,
     });
   }
 
@@ -58,34 +91,26 @@ class MethodChannelFlutterScreenguard extends FlutterScreenguardPlatform {
   ///
   /// [alignment] Alignment of the image, default Alignment.center
   ///
-  /// [timeAfterResume] (Android only) Time delayed for the view to stop displaying when going back
-  /// to the application (in milliseconds). Default = 1000ms
-  ///
-  /// function will throw warning when [timeAfterResume] bigger than 3000ms,
-  /// users have to wait for the application
-  /// to turn off the filter before going back to the main view, which is a very bad user
-  /// experiences.
-  ///
   /// Throws a [PlatformException] if there were technical problems on native side
   @override
-  Future<void> registerWithImage(
-      {required String uri,
-      required double width,
-      required double height,
-      Color? color = Colors.black,
-      Duration? timeAfterResume = const Duration(milliseconds: 1000),
-      Alignment? alignment,
-      double? top,
-      double? left,
-      double? bottom,
-      double? right}) async {
+  Future<void> registerWithImage({
+    required String uri,
+    required double width,
+    required double height,
+    Color? color = Colors.black,
+    Alignment? alignment,
+    double? top,
+    double? left,
+    double? bottom,
+    double? right,
+  }) async {
     final colorHex =
-        '#${color?.value.toRadixString(16).padLeft(8, '0').substring(2)}';
+        '#${color?.toARGB32().toRadixString(16).padLeft(8, '0').substring(2)}';
     final align = alignments.indexWhere(
       (element) => element == alignment,
     );
     await methodChannel
-        .invokeMethod<void>('registerWithImage', <String, dynamic>{
+        .invokeMethod<void>('activateShieldWithImage', <String, dynamic>{
       'uri': uri,
       'width': width.toString(),
       'height': height.toString(),
@@ -95,34 +120,23 @@ class MethodChannelFlutterScreenguard extends FlutterScreenguardPlatform {
       'bottom': bottom,
       'right': right,
       'color': colorHex,
-      'timeAfterResume': (timeAfterResume ?? const Duration(milliseconds: 1000))
-          .inMilliseconds,
     });
   }
 
   /// [iOS, Android] activate a screenshot blocking with a blurred effect view (iOS 13+, Android 8+)
   /// [radius] radius
   ///
-  /// [timeAfterResume] (Android only) Time delayed for the view to stop displaying when going back
-  /// to the application (in milliseconds). Default = 1000ms
-  ///
-  /// function will throw warning when [timeAfterResume] bigger than 3000ms,
-  /// users have to wait for the application to turn off the filter before going back
-  /// to the main view, which is a very bad user experiences.
-  ///
   /// Throws a [PlatformException] if there were technical problems on native side
   /// (e.g. lack of relevant hardware).
   @override
-  Future<void> registerWithBlurView(
-      {required num radius,
-      String? localImagePath,
-      Duration? timeAfterResume = const Duration(milliseconds: 1000)}) async {
+  Future<void> registerWithBlurView({
+    required num radius,
+    String? localImagePath,
+  }) async {
     await methodChannel
-        .invokeMethod<void>('registerWithBlurView', <String, dynamic>{
+        .invokeMethod<void>('activateShieldWithBlurView', <String, dynamic>{
       'radius': radius,
       'localImagePath': localImagePath,
-      'timeAfterResume': (timeAfterResume ?? const Duration(milliseconds: 1000))
-          .inMilliseconds,
     });
   }
 
@@ -130,6 +144,51 @@ class MethodChannelFlutterScreenguard extends FlutterScreenguardPlatform {
   /// Throws a [PlatformException] if there were technical problems on native side
   @override
   Future<void> unregister() async {
-    await methodChannel.invokeMethod<void>('unregister');
+    await methodChannel.invokeMethod<void>('deactivateShield');
+  }
+
+  /// [Android 5+] activate a screenshot blocking without any effect (blur, image, color)
+  @override
+  Future<void> registerWithoutEffect() async {
+    await methodChannel.invokeMethod<void>('activateShieldWithoutEffect');
+  }
+
+  @override
+  Future<List<Map<String, dynamic>>> getScreenGuardLogs({
+    required int maxCount,
+  }) async {
+    final List<dynamic>? logs = await methodChannel.invokeMethod<List<dynamic>>(
+      'getScreenGuardLogs',
+      {'maxCount': maxCount},
+    );
+    if (logs == null) {
+      return [];
+    }
+    return logs.cast<Map<String, dynamic>>();
+  }
+
+  @override
+  Future<void> initSettings({
+    bool? enableCapture = false,
+    bool? enableRecord = false,
+    bool? enableContentMultitask = false,
+    bool? displayOverlay = false,
+    bool? displayScreenguardOverlayAndroid = true,
+    int? timeAfterResume = 1000,
+    bool? getScreenshotPath = false,
+    int? limitCaptureEvtCount = 0,
+    bool? trackingLog = false,
+  }) async {
+    await methodChannel.invokeMethod('initSettings', {
+      'enableCapture': enableCapture,
+      'enableRecord': enableRecord,
+      'enableContentMultitask': enableContentMultitask,
+      'displayOverlay': displayOverlay,
+      'displayScreenguardOverlayAndroid': displayScreenguardOverlayAndroid,
+      'timeAfterResume': timeAfterResume,
+      'getScreenshotPath': getScreenshotPath,
+      'limitCaptureEvtCount': limitCaptureEvtCount,
+      'trackingLog': trackingLog,
+    });
   }
 }